:recycle: Move funding request letter retrieval functions to a service.
Why? Because it more closely matches the patterns I've used elsewhere.
e.g.
- a model file mapping to a database record
- a service that has retrieval functions that pull data from the database (in this case virtually).
