+++ src/api/controllers/admin/funding-request-letters-controller.ts
@@ -1,11 +1,11 @@
-import FundingRequestLetter from "@/models/funding-request-letter"
-
 import BaseController from "@/controllers/base-controller"
+
+import FundingRequestLettersService from "@/services/funding-request-letters-service"
 
 export default class FundingRequestLettersController extends BaseController {
   async listLetters() {
     if (this.request.params.fundingRequestId === undefined) {
-      const fundingRequestLetterMetadata = FundingRequestLetter.findAll()
+      const fundingRequestLetterMetadata = FundingRequestLettersService.all()
       return this.response.send(fundingRequestLetterMetadata)
     }
   }
